+++ unstructured/partition/pdf.py
@@ -2,6 +2,9 @@ import warnings
 from io import StringIO
 from typing import BinaryIO, List, Optional, cast
 
+from pdfminer.pdfpage import PDFPage, PDFTextExtractionNotAllowed
+from pdfminer.utils import open_filename
+
 from unstructured.documents.elements import Element, ElementMetadata, PageBreak
 from unstructured.logger import logger
 from unstructured.partition import _partition_via_api
@@ -99,17 +102,35 @@ def partition_pdf_or_image(
             out_template = None
 
         fallback_to_fast = False
+        fallback_to_hi_res = False
+
         detectron2_installed = dependency_exists("detectron2")
+        if is_image:
+            pdf_text_extractable = False
+        else:
+            pdf_text_extractable = is_pdf_text_extractable(filename=filename, file=file)
+            if file is not None:
+                file.seek(0)  # type: ignore
+
+        if not detectron2_installed and not pdf_text_extractable:
+            raise ValueError(
+                "detectron2 is not installed and the text of the PDF is not extractable. "
+                "To process this file, install detectron2 or remove copy protection from the PDF.",
+            )
+
+        if not pdf_text_extractable:
+            fallback_to_hi_res = strategy == "fast"
 
         if not detectron2_installed:
-            if is_image:
-                raise ValueError(
-                    "detectron2 is not installed. detectron2 is required for partioning images.",
-                )
-            else:
-                fallback_to_fast = True
+            fallback_to_fast = strategy == "hi_res"
 
-        if strategy == "hi_res" and not fallback_to_fast:
+        if (strategy == "hi_res" or fallback_to_hi_res) and not fallback_to_fast:
+            if strategy == "fast":
+                logger.warning(
+                    "PDF text is not extractable. Cannot use the fast partitioning "
+                    "strategy. Falling back to partitioning with the hi_res strategy.",
+                )
+
             # NOTE(robinson): Catches a UserWarning that occurs when detectron is called
             with warnings.catch_warnings():
                 warnings.simplefilter("ignore")
@@ -123,7 +144,7 @@ def partition_pdf_or_image(
                     ocr_languages=ocr_languages,
                 )
 
-        elif strategy == "fast" or fallback_to_fast:
+        elif (strategy == "fast" or fallback_to_fast) and not fallback_to_hi_res:
             if strategy == "hi_res":
                 logger.warning(
                     "detectron2 is not installed. Cannot use the hi_res partitioning "
@@ -234,9 +255,6 @@ def _partition_pdf_with_pdfminer(
 
     ref: https://github.com/pdfminer/pdfminer.six/blob/master/pdfminer/high_level.py
     """
-
-    from pdfminer.utils import open_filename
-
     exactly_one(filename=filename, file=file)
     if filename:
         with open_filename(filename, "rb") as fp:
@@ -270,14 +288,13 @@ def _process_pdfminer_pages(
     from pdfminer.converter import TextConverter
     from pdfminer.layout import LAParams
     from pdfminer.pdfinterp import PDFPageInterpreter, PDFResourceManager
-    from pdfminer.pdfpage import PDFPage
 
     rsrcmgr = PDFResourceManager(caching=False)
     laparams = LAParams()
 
     elements: List[Element] = []
 
-    for i, page in enumerate(PDFPage.get_pages(fp)):
+    for i, page in enumerate(PDFPage.get_pages(fp, check_extractable=True)):
         metadata = ElementMetadata(filename=filename, page_number=i + 1)
         with StringIO() as output_string:
             device = TextConverter(
@@ -298,3 +315,25 @@ def _process_pdfminer_pages(
             elements.append(PageBreak())
 
     return elements
+
+
+def is_pdf_text_extractable(filename: str = "", file: Optional[bytes] = None):
+    """Checks to see if the text from a PDF document is extractable. Sometimes the
+    text is not extractable due to PDF security settings."""
+    exactly_one(filename=filename, file=file)
+
+    def _fp_is_extractable(fp):
+        try:
+            next(PDFPage.get_pages(fp, check_extractable=True))
+            extractable = True
+        except PDFTextExtractionNotAllowed:
+            extractable = False
+        return extractable
+
+    if filename:
+        with open_filename(filename, "rb") as fp:
+            fp = cast(BinaryIO, fp)
+            return _fp_is_extractable(fp)
+    elif file:
+        fp = cast(BinaryIO, file)
+        return _fp_is_extractable(fp)